Добавил ряд исправлений.

+++ libraries/DS3231MDA/DS3231MDA.cpp
@@ -56,8 +56,8 @@ long DS3231MDA::time2long(uint16_t days, uint8_t h, uint8_t m, uint8_t s) {
 ///// ERIC'S ORIGINAL CODE FOLLOWS /////
 
 void DS3231MDA::getNow(byte &y, byte &m, byte &d, byte &hh, byte &mm, byte &ss) {
-  volatile unsigned int vAttempt=0;
-  unsigned int _maxcycles = microsecondsToClockCycles(65000);  // Пауза приблизительно в 65 мс   
+  volatile unsigned long vAttempt=0;
+  unsigned long  _maxcycles = microsecondsToClockCycles(65000);  // Пауза приблизительно в 65 мс   
   bool hasError=false;
   
  /****************************************************
